# docs/상품관리/03-design/01-openapi.yaml
openapi: 3.0.3
info:
  title: Product Management API
  description: |
    차세대 플랫폼 상품관리 API

    ## 주요 기능
    - 상품 CRUD (등록, 조회, 수정, 삭제)
    - 재고 관리 (입고, 출고, 조정)
    - 상태 관리 (DRAFT → PENDING → ON_SALE → SUSPENDED → ENDED)
    - 이미지 관리 (업로드, 삭제, 순서 변경)
    - 카테고리 관리

    ## 인증
    JWT Bearer Token 기반 인증

    ## Rate Limiting
    - POST /api/products: 10회/분
    - PUT /api/products/{id}: 30회/분
    - GET /api/products: 100회/분
    - GET /api/products/{id}: 200회/분
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://api-staging.example.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server

tags:
  - name: Products
    description: 상품 관리 API
  - name: Stock
    description: 재고 관리 API
  - name: Images
    description: 이미지 관리 API
  - name: Categories
    description: 카테고리 관리 API

paths:
  /api/products:
    post:
      tags:
        - Products
      summary: 신규 상품 등록
      description: 새로운 상품을 등록합니다. 상품코드는 자동 생성됩니다.
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
            examples:
              laptop:
                summary: 노트북 등록 예시
                value:
                  productName: 삼성 노트북 갤럭시북
                  categoryId: 123
                  price: 1500000
                  stock: 10
                  description: 최신 모델 노트북입니다.
                  options:
                    color: silver
                    storage: 512GB
                  images:
                    - https://cdn.example.com/temp/image1.jpg
                    - https://cdn.example.com/temp/image2.jpg
      responses:
        "201":
          description: 상품 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    get:
      tags:
        - Products
      summary: 상품 목록 조회
      description: 검색 조건에 맞는 상품 목록을 페이징하여 조회합니다.
      operationId: getProducts
      parameters:
        - name: search
          in: query
          description: 상품명 검색 (부분 일치)
          schema:
            type: string
            example: 노트북
        - name: categoryId
          in: query
          description: 카테고리 ID (하위 카테고리 포함)
          schema:
            type: integer
            format: int64
            example: 123
        - name: minPrice
          in: query
          description: 최소 가격
          schema:
            type: integer
            minimum: 0
            example: 1000000
        - name: maxPrice
          in: query
          description: 최대 가격
          schema:
            type: integer
            minimum: 0
            example: 2000000
        - name: status
          in: query
          description: 상품 상태
          schema:
            $ref: "#/components/schemas/ProductStatus"
        - name: sortBy
          in: query
          description: 정렬 기준
          schema:
            type: string
            enum: [createdAt, price, productName, viewCount]
            default: createdAt
        - name: sortOrder
          in: query
          description: 정렬 순서
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: 페이지 번호 (1부터 시작)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: 페이지 크기
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /api/products/{id}:
    get:
      tags:
        - Products
      summary: 상품 상세 조회
      description: 상품 코드로 상품 상세 정보를 조회합니다.
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Products
      summary: 상품 정보 수정
      description: 상품 정보를 수정합니다. 본인 상품 또는 관리자만 가능합니다.
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
            examples:
              updatePrice:
                summary: 가격 수정 예시
                value:
                  price: 1600000
              updateAll:
                summary: 전체 정보 수정 예시
                value:
                  productName: 삼성 노트북 갤럭시북 Pro
                  price: 1600000
                  stock: 15
                  description: 업그레이드된 최신 모델입니다.
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductUpdateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

    delete:
      tags:
        - Products
      summary: 상품 삭제 (논리 삭제)
      description: 상품을 논리 삭제합니다. 관리자만 가능합니다.
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  maxLength: 200
                  description: 삭제 사유
                  example: 단종 상품
      responses:
        "200":
          description: 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  productCode:
                    type: string
                  message:
                    type: string
                    example: 상품이 삭제되었습니다
                  deletedAt:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/products/{id}/stock:
    patch:
      tags:
        - Stock
      summary: 재고 수량 조정
      description: 상품의 재고 수량을 조정합니다 (입고/출고/조정).
      operationId: adjustStock
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentRequest"
            examples:
              stockIn:
                summary: 입고 예시
                value:
                  adjustmentType: IN
                  quantity: 50
                  reason: 신규 입고
              stockOut:
                summary: 출고 예시
                value:
                  adjustmentType: OUT
                  quantity: 10
                  reason: 판매 출고
      responses:
        "200":
          description: 조정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAdjustmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/products/{id}/status:
    patch:
      tags:
        - Products
      summary: 상품 상태 변경
      description: 상품의 상태를 변경합니다. 상태 전이 규칙을 준수해야 합니다.
      operationId: changeStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/ProductStatus"
      responses:
        "200":
          description: 상태 변경 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusChangeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/category:
    patch:
      tags:
        - Products
      summary: 카테고리 변경
      description: 상품의 카테고리를 변경합니다.
      operationId: changeCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - categoryId
              properties:
                categoryId:
                  type: integer
                  format: int64
                  description: 새 카테고리 ID
                  example: 456
      responses:
        "200":
          description: 카테고리 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  productCode:
                    type: string
                  previousCategoryId:
                    type: integer
                    format: int64
                  currentCategoryId:
                    type: integer
                    format: int64
                  changedAt:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/images:
    post:
      tags:
        - Images
      summary: 이미지 업로드
      description: 상품 이미지를 업로드합니다. 최대 10장까지 가능합니다.
      operationId: uploadImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 이미지 파일 (JPG, PNG, WebP, 최대 5MB)
                order:
                  type: integer
                  description: 이미지 순서
                  example: 1
                isThumbnail:
                  type: boolean
                  description: 대표 이미지 여부
                  default: false
      responses:
        "201":
          description: 업로드 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"

  /api/products/{id}/images/{imageId}:
    delete:
      tags:
        - Images
      summary: 이미지 삭제
      description: 상품 이미지를 삭제합니다.
      operationId: deleteImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
        - name: imageId
          in: path
          required: true
          description: 이미지 ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 이미지가 삭제되었습니다
                  deletedImageId:
                    type: integer
                    format: int64
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/images/order:
    put:
      tags:
        - Images
      summary: 이미지 순서 변경
      description: 상품 이미지의 순서를 변경합니다.
      operationId: reorderImages
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - imageOrders
              properties:
                imageOrders:
                  type: array
                  description: 이미지 ID와 순서 매핑
                  items:
                    type: object
                    required:
                      - imageId
                      - order
                    properties:
                      imageId:
                        type: integer
                        format: int64
                      order:
                        type: integer
            example:
              imageOrders:
                - imageId: 1
                  order: 3
                - imageId: 2
                  order: 1
                - imageId: 3
                  order: 2
      responses:
        "200":
          description: 순서 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 이미지 순서가 변경되었습니다
                  images:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductImage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/categories:
    get:
      tags:
        - Categories
      summary: 카테고리 목록 조회
      description: 전체 카테고리 목록을 계층 구조로 조회합니다.
      operationId: getCategories
      parameters:
        - name: level
          in: query
          description: 카테고리 레벨 (1, 2, 3)
          schema:
            type: integer
            enum: [1, 2, 3]
        - name: parentId
          in: query
          description: 부모 카테고리 ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token

  parameters:
    ProductId:
      name: id
      in: path
      required: true
      description: 상품 코드 (PRD-YYYYMMDD-NNNN)
      schema:
        type: string
        pattern: '^PRD-\d{8}-\d{4}$'
        example: PRD-20260206-0001

  schemas:
    ProductStatus:
      type: string
      enum:
        - DRAFT
        - PENDING
        - ON_SALE
        - SUSPENDED
        - ENDED
        - DELETED
      description: |
        상품 상태
        - DRAFT: 임시저장
        - PENDING: 판매대기
        - ON_SALE: 판매중
        - SUSPENDED: 판매중지
        - ENDED: 판매종료
        - DELETED: 삭제됨

    CreateProductRequest:
      type: object
      required:
        - productName
        - categoryId
        - price
        - stock
      properties:
        productName:
          type: string
          minLength: 2
          maxLength: 100
          description: 상품명
          example: 삼성 노트북 갤럭시북
        categoryId:
          type: integer
          format: int64
          description: 카테고리 ID
          example: 123
        price:
          type: integer
          minimum: 0
          multipleOf: 100
          description: 가격 (100원 단위)
          example: 1500000
        stock:
          type: integer
          minimum: 0
          description: 재고 수량
          example: 10
        description:
          type: string
          maxLength: 5000
          description: 상품 설명
          example: 최신 모델 노트북입니다.
        options:
          type: object
          additionalProperties: true
          description: 상품 옵션 (key-value)
          example:
            color: silver
            storage: 512GB
        images:
          type: array
          maxItems: 10
          items:
            type: string
            format: uri
          description: 이미지 URL 목록
          example:
            - https://cdn.example.com/temp/image1.jpg

    UpdateProductRequest:
      type: object
      properties:
        productName:
          type: string
          minLength: 2
          maxLength: 100
        price:
          type: integer
          minimum: 0
          multipleOf: 100
        stock:
          type: integer
          minimum: 0
        description:
          type: string
          maxLength: 5000
        options:
          type: object
          additionalProperties: true

    ProductResponse:
      type: object
      properties:
        productCode:
          type: string
          example: PRD-20260206-0001
        productName:
          type: string
          example: 삼성 노트북 갤럭시북
        categoryId:
          type: integer
          format: int64
          example: 123
        price:
          type: integer
          example: 1500000
        stock:
          type: integer
          example: 10
        status:
          $ref: "#/components/schemas/ProductStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"

    ProductDetailResponse:
      allOf:
        - $ref: "#/components/schemas/ProductResponse"
        - type: object
          properties:
            categoryPath:
              type: string
              example: 전자제품 > 노트북 > 게이밍
            description:
              type: string
            options:
              type: object
              additionalProperties: true
            seller:
              $ref: "#/components/schemas/Seller"
            viewCount:
              type: integer
              example: 1523

    ProductUpdateResponse:
      allOf:
        - $ref: "#/components/schemas/ProductResponse"
        - type: object
          properties:
            version:
              type: integer
              example: 2
            changeHistory:
              $ref: "#/components/schemas/ChangeHistory"

    ProductListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"
        page:
          $ref: "#/components/schemas/PageInfo"

    ProductSummary:
      type: object
      properties:
        productCode:
          type: string
        productName:
          type: string
        price:
          type: integer
        stock:
          type: integer
        status:
          $ref: "#/components/schemas/ProductStatus"
        thumbnailUrl:
          type: string
          format: uri
        categoryName:
          type: string
        createdAt:
          type: string
          format: date-time

    ProductImage:
      type: object
      properties:
        imageId:
          type: integer
          format: int64
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        mediumUrl:
          type: string
          format: uri
        order:
          type: integer
        isThumbnail:
          type: boolean

    StockAdjustmentRequest:
      type: object
      required:
        - adjustmentType
        - quantity
        - reason
      properties:
        adjustmentType:
          type: string
          enum: [IN, OUT, ADJUST]
          description: 조정 유형 (입고/출고/조정)
        quantity:
          type: integer
          minimum: 1
          description: 조정 수량
        reason:
          type: string
          maxLength: 200
          description: 조정 사유

    StockAdjustmentResponse:
      type: object
      properties:
        productCode:
          type: string
        previousStock:
          type: integer
        adjustedStock:
          type: integer
        stockHistory:
          $ref: "#/components/schemas/StockHistory"

    StockHistory:
      type: object
      properties:
        historyId:
          type: integer
          format: int64
        adjustmentType:
          type: string
        quantity:
          type: integer
        reason:
          type: string
        adjustedBy:
          type: string
        adjustedAt:
          type: string
          format: date-time

    StatusChangeResponse:
      type: object
      properties:
        productCode:
          type: string
        previousStatus:
          $ref: "#/components/schemas/ProductStatus"
        currentStatus:
          $ref: "#/components/schemas/ProductStatus"
        statusChangedAt:
          type: string
          format: date-time

    ChangeHistory:
      type: object
      properties:
        historyId:
          type: integer
          format: int64
        changedFields:
          type: array
          items:
            type: string
        changedBy:
          type: string

    Seller:
      type: object
      properties:
        sellerId:
          type: integer
          format: int64
        sellerName:
          type: string

    Category:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        parentId:
          type: integer
          format: int64
          nullable: true
        level:
          type: integer
          enum: [1, 2, 3]
        productCount:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"

    PageInfo:
      type: object
      properties:
        number:
          type: integer
          description: 현재 페이지 번호
        size:
          type: integer
          description: 페이지 크기
        totalElements:
          type: integer
          format: int64
          description: 전체 요소 수
        totalPages:
          type: integer
          description: 전체 페이지 수

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
        field:
          type: string
          description: 에러 발생 필드 (선택적)
        timestamp:
          type: string
          format: date-time
          description: 에러 발생 시각

  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingField:
              summary: 필수 필드 누락
              value:
                error: E004
                message: 필수 정보가 누락되었습니다
                field: price
                timestamp: 2026-02-06T11:44:00Z
            invalidPrice:
              summary: 가격 단위 오류
              value:
                error: INVALID_PRICE
                message: 가격은 100원 단위여야 합니다
                field: price
                timestamp: 2026-02-06T11:44:00Z

    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: UNAUTHORIZED
            message: 인증이 필요합니다
            timestamp: 2026-02-06T11:44:00Z

    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: E003
            message: 권한이 없습니다
            timestamp: 2026-02-06T11:44:00Z

    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: E002
            message: 존재하지 않는 상품입니다
            timestamp: 2026-02-06T11:44:00Z

    Conflict:
      description: 충돌 발생
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            versionConflict:
              summary: 낙관적 락 충돌
              value:
                error: VERSION_CONFLICT
                message: 다른 사용자가 수정 중입니다
                timestamp: 2026-02-06T11:44:00Z
            insufficientStock:
              summary: 재고 부족
              value:
                error: E001
                message: 재고가 부족합니다
                timestamp: 2026-02-06T11:44:00Z

    PayloadTooLarge:
      description: 요청 크기 초과
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: E006
            message: 이미지 크기가 초과되었습니다
            timestamp: 2026-02-06T11:44:00Z

    TooManyRequests:
      description: 요청 한도 초과
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: RATE_LIMIT_EXCEEDED
            message: 요청 한도를 초과했습니다
            timestamp: 2026-02-06T11:44:00Z